backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.4
pydantic_core==2.46.4
python-multipart==0.0.27
httpx==0.28.1
openai==2.35.1
PyJWT==2.12.1
PyMySQL==1.1.3
python-docx==1.2.0
docx2pdf==0.1.8
pywin32==311
xhtml2pdf==0.2.17
reportlab==4.5.0
lxml==6.1.0
Pillow==12.2.0
starlette==1.0.0
anyio==4.13.0
